Move plot to separate components
